refactor(scripts): split V1 mapping into its own module
structural-auditor flagged import-v1-representation.mjs at 516 lines
(>400, forbidden). Extract the V1 typedefs and mapCompanyFromV1 /
mapDeclarationFromV1 / computeReferencePeriodStart into a sibling
import-v1-representation-mapping.mjs, re-exported from the entrypoint so
the public import path (and tu-dev's tests) are unaffected. Both files
now under 400 lines. No behavior change — full unit + integration suite
re-verified green.
